+++ app/server/request-context.ts
@@ -17,30 +17,3 @@ export function getRequestContext() {
 export function requestContextStorageHandler(): RequestHandler {
   return async (ctx) => requestContextStorage.run(ctx, () => ctx.next());
 }
-
-//
-// allow manipulating response headers via context
-// e.g. for set-cookie, cache-control, ...
-//
-
-declare module "@hattip/compose" {
-  interface RequestContextExtensions {
-    responseHeaders: Headers;
-  }
-}
-
-export function responseHeadersContextHandler(): RequestHandler {
-  return async (ctx) => {
-    ctx.responseHeaders = new Headers();
-    const res = await ctx.next();
-    ctx.responseHeaders.forEach((v, k) => {
-      res.headers.set(k, v);
-    });
-    return res;
-  };
-}
-
-// full cache only on CDN (cf. https://vercel.com/docs/concepts/functions/serverless-functions/edge-caching)
-export function cacheResponse(headers: Headers) {
-  headers.set("cache-control", "public, max-age=0, s-max-age=31536000");
-}
